move travis/mailer to the email addon, add an optional setup method for addons
