+++ lib/travis/addons/flowdock.rb
@@ -2,12 +2,8 @@ module Travis
   module Addons
     module Flowdock
       autoload :EventHandler, 'travis/addons/flowdock/event_handler'
+      autoload :Instruments,  'travis/addons/flowdock/instruments'
       autoload :Task,         'travis/addons/flowdock/task'
-
-      module Instruments
-        autoload :EventHandler, 'travis/addons/flowdock/instruments'
-        autoload :Task,         'travis/addons/flowdock/instruments'
-      end
     end
   end
 end
