+++ lib/travis/addons/irc.rb
@@ -3,12 +3,8 @@ module Travis
     module Irc
       autoload :Client,       'travis/addons/irc/client'
       autoload :EventHandler, 'travis/addons/irc/event_handler'
+      autoload :Instruments,  'travis/addons/irc/instruments'
       autoload :Task,         'travis/addons/irc/task'
-
-      module Instruments
-        autoload :EventHandler, 'travis/addons/irc/instruments'
-        autoload :Task,         'travis/addons/irc/instruments'
-      end
     end
   end
 end
