+++ lib/travis/addons/pusher.rb
@@ -2,12 +2,8 @@ module Travis
   module Addons
     module Pusher
       autoload :EventHandler, 'travis/addons/pusher/event_handler'
+      autoload :Instruments,  'travis/addons/pusher/instruments'
       autoload :Task,         'travis/addons/pusher/task'
-
-      module Instruments
-        autoload :EventHandler, 'travis/addons/pusher/instruments'
-        autoload :Task,         'travis/addons/pusher/instruments'
-      end
     end
   end
 end
