+++ spec/travis/addons/email/mailer/helpers_spec.rb
@@ -1,7 +1,7 @@
 require 'spec_helper'
 
-describe Travis::Mailer::Helper::Build do
-  include Travis::Mailer::Helper::Build, Travis::Testing::Stubs
+describe Travis::Addons::Email::Mailer::Helpers do
+  include Travis::Addons::Email::Mailer::Helpers, Travis::Testing::Stubs
 
   it '#title returns title for the build' do
     title(repository).should == 'Build Update for svenfuchs/minimal'
